Progress commit: BitMex service

+++ APISandbox/Models/BitMexHistoricalOrderResult.cs
@@ -13,16 +13,16 @@ namespace APISandbox.Models
     }
         public class BitMexHistoricalOrderResultOrder
     {
-        [JsonPropertyName("account")] public double Account { get; set; }
-        [JsonPropertyName("avgPx")] public double AvgPx { get; set; }
-        [JsonPropertyName("cumQty")] public double CumQty { get; set; }
+        [JsonPropertyName("account")] public float Account { get; set; }
+        [JsonPropertyName("avgPx")] public float AvgPx { get; set; }
+        [JsonPropertyName("cumQty")] public float CumQty { get; set; }
         [JsonPropertyName("currency")] public string Currency { get; set; }
-        [JsonPropertyName("leavesQty")] public double LeavesQty { get; set; }
+        [JsonPropertyName("leavesQty")] public float LeavesQty { get; set; }
         [JsonPropertyName("ordStatus")] public string OrdStatus { get; set; }
         [JsonPropertyName("ordType")] public string OrdType { get; set; }
         [JsonPropertyName("orderID")] public string OrderID { get; set; }
-        [JsonPropertyName("orderQty")] public double OrderQty { get; set; }
-        [JsonPropertyName("price")] public double Price { get; set; }
+        [JsonPropertyName("orderQty")] public float OrderQty { get; set; }
+        [JsonPropertyName("price")] public float Price { get; set; }
         [JsonPropertyName("side")] public string Side { get; set; }
         [JsonPropertyName("symbol")] public string Symbol { get; set; }
         [JsonPropertyName("text")] public string Text { get; set; }
